将原来的颜色选择框拉进来到画笔面板类中，这样方便等会的各种变换

+++ StoryBox/Models/SBPen.h
@@ -14,5 +14,5 @@
 @interface SBPen : NSObject
 
 @property (nonatomic) int color;
-@property (nonatomic) float radius;
+@property (nonatomic) unsigned int radius;
 @end
